+++ Solution Items/RELEASENOTES.md
@@ -1,5 +1,11 @@
 # RELEASE NOTES
 
+**1.1.0**
+
+- Replaced current file header with a Yaml Front Matter (breaking!!)
+- Removed MenuBuilder (this can be done via templating now)
+- Available variables for templating changed, please see documentation
+
 **1.0.1**
 
 - Added support of Date